requirements.txt
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1733235340728/work
bandit==1.8.3
billiard==4.2.1
celery==5.5.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1754063596306/work
coverage==7.10.2
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
Django==5.2.5
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
environs==9.5.0
et_xmlfile==2.0.0
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1752722674086/work
grpcio==1.74.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1751493861733/work
kombu==5.5.4
markdown-it-py==3.0.0
marshmallow==3.22.0
matplotlib==3.10.5
mdurl==0.1.2
milvus-lite==2.4.12
munkres==1.1.4
mysql-connector-python @ file:///home/conda/feedstock_root/build_artifacts/mysql-connector-python_1745443727395/work/mysql-connector-python
mysqlclient==2.2.7
numpy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_numpy_1753401560/work/dist/numpy-2.3.2-cp311-cp311-linux_x86_64.whl#sha256=469440884ff65cdd5601d2ff9e1c5b9bcc2d8176c37e37ec78a0666703a54157
openpyxl==3.1.5
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1752081700120/work
pbr==6.1.1
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1751481992986/work
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==6.31.1
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pymilvus==2.4.8
PyMuPDF==1.24.14
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1753873557/work
PySide6==6.9.1
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==14.1.0
shiboken6==6.9.1
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sqlparse==0.5.3
stevedore==5.4.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1748003302865/work
tqdm==4.67.1
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
ujson==5.10.0
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1736692517942/work
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
